Cleaned and made modular; removed old modules and replaced them
