--- CTestTestfile.cmake
@@ -1,8 +1,0 @@
-# CMake generated Testfile for 
-# Source directory: /home/ubuntu/workspace
-# Build directory: /home/ubuntu/workspace
-# 
-# This file includes the relevant testing commands required for 
-# testing this directory and lists subdirectories to be tested as well.
-SUBDIRS(src/lib/modules)
-SUBDIRS(src/tests)
