--- README.md
@@ -1,13 +1,0 @@
-# OpenRPG
-A handy dandy helper program for playing D&amp;D! (This is just a concept and will need to be expanded on)
-
-# Building
-Currently we require `cmake` (at least v3.0) and `make`
-
-Create a build directory `$ mkdir build &&  cd build`  
-Run cmake `$ cmake ..`  
-Run make `$ make`  
-Run make check `$ make check`  
-
-# Installing
-After building run make install `$ make install`
