--- Testing/Temporary/LastTest.log
@@ -1,33 +1,0 @@
-Start testing: Jan 13 17:32 UTC
-----------------------------------------------------------
-1/2 Testing: generator-test-dwarf-male
-1/2 Test: generator-test-dwarf-male
-Command: "/home/ubuntu/workspace/src/tests/generator-test" "dwarf" "male"
-Directory: /home/ubuntu/workspace/src/tests
-"generator-test-dwarf-male" start time: Jan 13 17:32 UTC
-Output:
-----------------------------------------------------------
-<end of output>
-Test time =   0.77 sec
-----------------------------------------------------------
-Test Passed.
-"generator-test-dwarf-male" end time: Jan 13 17:32 UTC
-"generator-test-dwarf-male" time elapsed: 00:00:00
-----------------------------------------------------------
-
-2/2 Testing: generator-test-dwarf-female
-2/2 Test: generator-test-dwarf-female
-Command: "/home/ubuntu/workspace/src/tests/generator-test" "dwarf" "female"
-Directory: /home/ubuntu/workspace/src/tests
-"generator-test-dwarf-female" start time: Jan 13 17:32 UTC
-Output:
-----------------------------------------------------------
-<end of output>
-Test time =   0.74 sec
-----------------------------------------------------------
-Test Passed.
-"generator-test-dwarf-female" end time: Jan 13 17:32 UTC
-"generator-test-dwarf-female" time elapsed: 00:00:00
-----------------------------------------------------------
-
-End testing: Jan 13 17:32 UTC
