--- Testing/Temporary/LastTestsFailed.log
@@ -1,2 +1,0 @@
-1:generator-test-dwarf-male
-2:generator-test-dwarf-female
