--- src/assets/names/dwarf/last
@@ -1,15 +1,0 @@
-Balderk
-Battlehammer
-Brawnanvil
-Dankil
-Fireforge
-Frostbeard
-Gorunn
-Holderhek
-Ironfist
-Loderr
-Lutgehr
-Rumnaheim
-Strakeln
-Torunn
-Ungart
