--- src/assets/names/dwarf/male
@@ -1,30 +1,0 @@
-Adrik
-Alberich
-Baern
-Barendd
-Brottor
-Bruenor
-Dain
-Darrak
-Delg
-Eberk
-Einkil
-Fargrim
-Flint
-Gardain
-Harbek
-Kildrak
-Morgran
-Orsik
-Oskar
-Rangrim
-Rurik
-Taklinn
-Thoradin
-Thorin
-Tordek
-Traubon
-Travok
-Ulfgar
-Veit
-Vondal
