+++ src/lib/config.h
@@ -1,6 +1,6 @@
 // the configured options and settings for Tutorial
 #define OpenRPG_VERSION_MAJOR 0
 #define OpenRPG_VERSION_MINOR 0
-#define OpenRPG_VERSION_TWEAK 2
-#define OpenRPG_VERSION_SUFFIX "dev"
+#define OpenRPG_VERSION_TWEAK 3
+#define OpenRPG_VERSION_SUFFIX "shiny"
 #define ASSET_LOC "../assets"
