+++ src/lib/modules/roll/roll.h
@@ -2,7 +2,7 @@
 #define GENERATE_H
 
 int nameGenerator(std::string gender, std::string race);
-int parse(std::string in);
+
 int roll(int faces);
 
 #endif
