+++ src/lib/utils.h
@@ -17,6 +17,5 @@ std::string rightpad(std::string str, int len = 0, char ch = ' ');
 std::istream& safeGetline(std::istream& is, std::string& t);
 bool print_file(std::string type);
 int parse(std::string in);
-int roll();
 
 #endif /* SRC_UTILS_H_ */
