--- src/tests/CTestTestfile.cmake
@@ -1,8 +1,0 @@
-# CMake generated Testfile for 
-# Source directory: /home/ubuntu/workspace/src/tests
-# Build directory: /home/ubuntu/workspace/src/tests
-# 
-# This file includes the relevant testing commands required for 
-# testing this directory and lists subdirectories to be tested as well.
-ADD_TEST(generator-test-dwarf-male "generator-test" "dwarf" "male")
-ADD_TEST(generator-test-dwarf-female "generator-test" "dwarf" "female")
