--- src/tests/assets/banners/welcome_mat1
@@ -1,15 +1,0 @@
-
-                          Welcome to:             
-    /:\                                                  (""")   
-    |:|                                                   III    
-    |:|                                                 __III__  
-    |:|                                               /:-.___,-:\
-    |:|        _____             _____ _____ _____    \]  |:|  [/
-    |:|       |     |___ ___ ___| __  |  _  |   __|       |:|    
-/]  |:|  [\   |  |  | . | -_|   |    -|   __|  |  |       |:|    
-\\:-''''-:/   |_____|  _|___|_|_|__|__|__|  |_____|       |:|    
-  ""III""           |_|                                   |:|    
-    III                                                   |:|    
-   (___)                                                  \:/    
-                            OpenRPG                               
-
