--- src/tests/assets/names/dwarf/female
@@ -1,23 +1,0 @@
-Amber
-Artin
-Audhild
-Bardryn
-Dagnal
-Diesa
-Eldeth
-Falkrunn
-Finellen
-Gunnloda
-Gurdis
-Helja
-Hlin
-Kathra
-Kristryd
-Ilde
-Liftrasa
-Mardred
-Riswynn
-Sannl
-Torbera
-Torgga
-Vistra
